model/requirement.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
numpy==1.24.3
praw==7.5.0
python-dotenv==0.19.2
scikit-learn==1.0.1